+++ src/repository/meal_repository.rs
@@ -8,7 +8,8 @@ use crate::{
     schema::meals,
 };
 
-use super::{connection_psql::establish_connection, paginator::Paginate};
+use super::paginator::Paginate;
+use crate::connection::connection_psql::establish_connection;
 
 pub fn ingest_meal<T>(new_meal: T) -> Result<usize, Error>
 where
