+++ src/repository/mod.rs
@@ -1,11 +1,9 @@
-pub mod user_repository;
-pub mod baby_repository;
-pub mod connection_redis;
-pub mod session_repository;
-pub mod connection_psql;
-pub mod association_repository;
-pub mod meal_repository;
-pub mod dream_repository;
-pub mod weight_repository;
-pub mod paginator;
 pub mod admin_repository;
+pub mod association_repository;
+pub mod baby_repository;
+pub mod dream_repository;
+pub mod meal_repository;
+pub mod paginator;
+pub mod session_repository;
+pub mod user_repository;
+pub mod weight_repository;
